release 0.7.2

+++ HyperchainCore/node/UdpRecvDataHandler.hpp
@@ -1,4 +1,4 @@
-/*Copyright 2016-2020 hyperchain.net (Hyperchain)
+/*Copyright 2016-2021 hyperchain.net (Hyperchain)
 
 Distributed under the MIT software license, see the accompanying
 file COPYING or?https://opensource.org/licenses/MIT.
@@ -79,7 +79,7 @@ public:
         }
 
 
-        _dispatcher.dispatch(taskbuf->c_str(), taskbuf->size(), string(ip), port);
+        _dispatcher.dispatch(taskbuf->c_str(), (int)taskbuf->size(), string(ip), port);
 
         return true;
     }
